deleted thisday in tests, added input func to thisday
