+++ tests/test_thisday.py
@@ -1,5 +1,5 @@
 import pytest
-from thisday import thisday
+from thisday import thisday,ProcessInput
 
 class Tests:
     # function tests go here
@@ -19,4 +19,11 @@ class Tests:
         expected = True
         actual = True
         assert actual == expected, 'Expected True to equal True'
+
+    def test_input():
+        assert ProcessInput("film-tv")=="film-tv","film-tv doesn't run correctly"
+        assert ProcessInput("history")=="history","history doesn't run correctly"
+        assert ProcessInput("sport")=="sport","sport doesn't run correctly"
+        assert ProcessInput("music")=="music","music doesn't run correctly"
+        assert ProcessInput("foo")==0,"other cases doesn't run correctly"
