--- tests/thisday.py
@@ -1,7 +1,0 @@
-def ProcessInput(inputString):
-    validInput=['film-tv','history','sport','music']
-    if inputString in validInput:
-        return inputString
-    else:
-        print("Incorrect Input, please try again.")
-
